requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2
instagrapi
openai==1.6.1
pillow==10.1.0
python-multipart==0.0.6
google-api-python-client==2.111.0
yt-dlp==2023.11.16
apscheduler==3.10.4
jinja2==3.1.2
requests==2.31.0
apify-client==1.6.0
celery==5.3.6
redis==5.0.1
boto3==1.34.14
